requirements.txt
alembic==1.4.1
attrs==20.3.0
azure-core==1.12.0
azure-storage-blob==12.8.0
boto3==1.17.30
botocore==1.20.30
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
configparser==5.0.2
connexion==2.4.0
cryptography==3.4.6
databricks-cli==0.14.3
docker==4.4.4
ecdsa==0.14.1
elasticsearch==7.11.0
entrypoints==0.3
etcd3==0.12.0
Flask==1.1.2
Flask-Cors==3.0.10
gevent==21.1.2
gitdb==4.0.5
GitPython==3.1.14
gorilla==0.3.0
greenlet==1.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
minio==6.0.0
mlflow==1.12.1
msrest==0.6.21
numpy==1.20.1
oauthlib==3.1.0
openapi-schema-validator==0.1.4
openapi-spec-validator==0.3.0
pandas==1.2.3
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.15.6
pyasn1==0.4.8
pycparser==2.20
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
python-jose==3.2.0
python-keycloak-client==0.2.3
pytz==2021.1
PyYAML==5.2
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.4
scikit-learn==0.23.2
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==3.0.5
SQLAlchemy==1.4.1
sqlparse==0.4.1
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==7.0.0
threadpoolctl==2.1.0
typing==3.7.4.3
urllib3==1.26.4
waitress==2.0.0
websocket-client==0.58.0
Werkzeug==0.16.1
zope.event==4.5.0
zope.interface==5.2.0
